Refactored the core c++ logic for searhc and also refactored all the runtime test cases

+++ src/core/search_service.h
@@ -28,7 +28,7 @@ private:
     InvertedIndex idx_;
 
     // docId -> document length (token count)
-    std::unordered_map<int, int> doc_len_;
+    std::unordered_map<int, size_t> doc_len_;
     std::unordered_map<int, std::string> doc_text_;
 
     // number of documents indexed
